Fix Modulo Profesores

+++ app/Distinciones.php
@@ -23,16 +23,13 @@ class Distinciones extends Model
         $periodo = $request->get('Periodo');
         $tipoDistincion = $request->get('Nombre_Distincion');
         
+        $NombreArchivo="";
+        $RutaArchivo="";
+
         if($tipoDistincion=="0"){
             $tipoDistincion = $request->get('OpcionOtro');
         }
-        
-        try {
-            mkdir(public_path().'/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/', 0700);
-        } catch (\Throwable $th) {
-            //throw $th;
-        }
-        
+
         $profesor = new Profesores();
         $profesor = $profesor->ObtenerNombreProfesor($request->get('idProfesor'));
         $inicialesProf = "";
@@ -43,20 +40,32 @@ class Distinciones extends Model
         $inicialesProf = $inicialesProf . substr($profesor[0]->Apellido_P, 0, 1);
         $inicialesProf = $inicialesProf . substr($profesor[0]->Apellido_M, 0, 1);
         
-        $extension = $archivo->getClientOriginalExtension();
-        $fileName = $inicialesProf."-".$tipoDistincion."-".$periodo.'-'.$anio.'.'.$extension;
-        $tmpPath = $archivo;
-        $newPath = public_path().'/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName;
-        move_uploaded_file($tmpPath,$newPath);
+        if($archivo!=null){
+            try {
+                mkdir(public_path().'/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/', 0700);
+            } catch (\Throwable $th) {
+                //throw $th;
+            }
+    
+            $extension = $archivo->getClientOriginalExtension();
+            $fileName = $inicialesProf."-".$tipoDistincion."-".$periodo.'-'.$anio.'.'.$extension;
+            $tmpPath = $archivo;
+            $newPath = public_path().'/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName;
+            move_uploaded_file($tmpPath,$newPath);    
+            
+            $NombreArchivo=$fileName;
+            $RutaArchivo='/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName;
+        }
 
+        
         $datos = new Distinciones ([
             'Id_Profesor'=> $request->get('idProfesor'),
             'Nombre_Distincion'=> $tipoDistincion,
             'Periodo'=> $request->get('Periodo'),
             'Anio'=> $request->get('Anio'),
             'Descripcion'=> $request->get('Descripcion'),
-            'Ruta_Archivo'=> '/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName,
-            'NombreArchivo'=> $fileName
+            'Ruta_Archivo'=> $RutaArchivo,
+            'NombreArchivo'=> $NombreArchivo
         ]); 
         $datos->save();
 
@@ -67,8 +76,10 @@ class Distinciones extends Model
         $ids = $request->get('Ids');
         for ($i=0; $i < count($ids); $i++) { 
             $Distincion = Distinciones::find($ids[$i]);
-            $file = glob(public_path().$Distincion->Ruta_Archivo);
-            unlink($file[0]); 
+            if ($Distincion->Ruta_Archivo!=null) {
+                $file = glob(public_path().$Distincion->Ruta_Archivo);
+                unlink($file[0]); 
+            }
             $Distincion->delete();
         }
 
@@ -81,6 +92,9 @@ class Distinciones extends Model
         $periodo = $request->get('Periodo');
         $tipoDistincion = $request->get('Nombre_Distincion');
         
+        $NombreArchivo="";
+        $RutaArchivo="";
+
         $idRegistro = $request->get('IdRegistro');
 
         $Distincion = Distinciones::find($idRegistro);
@@ -102,6 +116,11 @@ class Distinciones extends Model
         $filename="";
         if($archivo!=null){
             try {
+                mkdir(public_path().'/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/', 0700);
+            } catch (\Throwable $th) {
+                //throw $th;
+            }
+            try {
                 $file = glob(public_path().$Distincion->Ruta_Archivo);
                 unlink($file[0]);
             }catch(\Throwable $th){} 
@@ -111,11 +130,18 @@ class Distinciones extends Model
             $tmpPath = $archivo;
             $newPath = public_path().'/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName;
             move_uploaded_file($tmpPath,$newPath);
+
+            $NombreArchivo=$fileName;
+            $RutaArchivo='/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName;
         }else{
-            $extension = "pdf";
-            $fileName = $inicialesProf."-".$tipoDistincion."-".$periodo.'-'.$anio.'.'.$extension;
-            $newPath = public_path().'/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName;
-            rename(public_path().$Distincion->Ruta_Archivo,$newPath);
+            if ($Distincion->Ruta_Archivo!=null) {
+                $extension = "pdf";
+                $fileName = $inicialesProf."-".$tipoDistincion."-".$periodo.'-'.$anio.'.'.$extension;
+                $newPath = public_path().'/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName;
+                rename(public_path().$Distincion->Ruta_Archivo,$newPath);
+                $NombreArchivo=$fileName;
+                $RutaArchivo='/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName;
+            }
         }
 
         $Distincion->Nombre_Distincion = $tipoDistincion;
@@ -123,8 +149,8 @@ class Distinciones extends Model
         $Distincion->Anio = $request->get('Anio');
         $Distincion->Periodo = $request->get('Periodo');
         $Distincion->Descripcion = $request->get('Descripcion');
-        $Distincion->Ruta_Archivo = '/storage/Documentos/Profesores/Distincion/'.$request->get('idProfesor').'/'.$fileName;
-        $Distincion->NombreArchivo = $fileName;
+        $Distincion->Ruta_Archivo = $RutaArchivo;
+        $Distincion->NombreArchivo = $NombreArchivo;
         
         $Distincion->save();
